feat: changed amarex components back to regular components in addons folder
