+++ addons/addons-store/gettersAddons.js
@@ -1,11 +1,11 @@
 import AbimoHandler from "../abimoHandler/components/AbimoHandler.vue";
-import BaseMaps from "../baseMaps/components/BaseMapsAmarex.vue";
+import BaseMaps from "../baseMaps/components/BaseMaps.vue";
 import FileImporter from "../fileImporter/components/FileImporter.vue";
 import ProjectDownloader from "../projectDownloader/components/ProjectDownloader.vue";
 import ProjectStarter from "../projectStarter/components/ProjectStarter.vue";
 import ProjectUploader from "../projectUploader/components/ProjectUploader.vue";
 import ReportPrinter from "../reportPrinter/components/ReportPrinter.vue";
-import ThemeMaps from "../themeMaps/components/ThemeMapsAmarex.vue";
+import ThemeMaps from "../themeMaps/components/ThemeMaps.vue";
 
 const getters = {
   componentMap: () => {
